Fix `Promise.prototype.finally` polyfill loading in older browsers
The sidebar application failed to load in Safari 10 due to the ES 2018
polyfill bundle, which contains a polyfill for
`Promise.prototype.finally`, not being loaded in that browser.

The issue was that 'es2018' was not listed in the polyfill bundles to
build in `gulpfile.js` or in the list of bundles required by the sidebar
app in `boot.js`.

This commit fixes these issues and makes the issue less likely to occur
in future by:

 - Removing the need to manually list each polyfill bundle in
   `gulpfile.js`. Instead the `src/shared/polyfills` dir is listed and
   every non-`index.js` file is used as the root of a bundle.
 - Factoring out the polyfills common to the sidebar and annotator app
   into a variable and including "es2018" there